Initial First-Party Typings
This is being type checked with JSDoc initially, then I will be able to move the codebase to TypeScript later on, once things are safely documented and everything is ship-shape.

Listening back to Contraire de la Chanson, and wow it really is an absolutely outstanding album, for me at least haha.