Private Temp Disable
This is because of the JSDoc linting of types for class fields, when using non-ES6 class implementations. This seems to be a bug with TypeScript, unless it's intended behavior for some reason. Essentially it's flagging the use of assigning to fields when inside the constructor, while in regular TS this is perfectly fine. Marking the function with 'class' or 'constructor' still doesn't seem to fix it unfortunately.