DOM Parsing Types
Further working on the JSDoc typings for the existing library code. Having these will allow me to for the most part, safely refactor the code, alongside validating that the unit tests are still passing properly as well. Having first-party type definitions are an awesome thing too, but I think it's even better for the project to be able to typecheck itself, as you can catch so many goofy errors with it along the way (personal experience talking ha).